web/requirements.txt
flask
flask_sqlalchemy
gunicorn
psycopg2
pyyaml
